Reviewed files according to the feedback
